refactor: move out calendar name input and select color
